Move gamma to monitor
