--- gamma.go
@@ -1,53 +1,0 @@
-package glfw3
-
-//#include <GLFW/glfw3.h>
-//unsigned int GetGammaAtIndex(unsigned short *color, int i);
-//void SetGammaAtIndex(unsigned short *color, int i, unsigned short value);
-import "C"
-
-//GammaRamp describes the gamma ramp for a monitor.
-type GammaRamp struct {
-	Red   []uint16 //A slice of value describing the response of the red channel.
-	Green []uint16 //A slice of value describing the response of the green channel.
-	Blue  []uint16 //A slice of value describing the response of the blue channel.
-}
-
-//SetGamma generates a 256-element gamma ramp from the specified exponent and then calls
-//SetGamma with it.
-func (m *Monitor) SetGamma(gamma float32) {
-	C.glfwSetGamma(m.data, C.float(gamma))
-}
-
-//GetGammaRamp retrieves the current gamma ramp of the monitor.
-func (m *Monitor) GetGammaRamp() *GammaRamp {
-	var ramp GammaRamp
-
-	rampC := C.glfwGetGammaRamp(m.data)
-	length := int(rampC.size)
-	ramp.Red = make([]uint16, length)
-	ramp.Green = make([]uint16, length)
-	ramp.Blue = make([]uint16, length)
-
-	for i := 0; i < length; i++ {
-		ramp.Red[i] = uint16(C.GetGammaAtIndex(rampC.red, C.int(i)))
-		ramp.Green[i] = uint16(C.GetGammaAtIndex(rampC.green, C.int(i)))
-		ramp.Blue[i] = uint16(C.GetGammaAtIndex(rampC.blue, C.int(i)))
-	}
-
-	return &ramp
-}
-
-//SetGammaRamp sets the current gamma ramp for the monitor.
-func (m *Monitor) SetGammaRamp(ramp *GammaRamp) {
-	var rampC C.GLFWgammaramp
-
-	length := len(ramp.Red)
-
-	for i := 0; i < length; i++ {
-		C.SetGammaAtIndex(rampC.red, C.int(i), C.ushort(ramp.Red[i]))
-		C.SetGammaAtIndex(rampC.green, C.int(i), C.ushort(ramp.Green[i]))
-		C.SetGammaAtIndex(rampC.blue, C.int(i), C.ushort(ramp.Blue[i]))
-	}
-
-	C.glfwSetGammaRamp(m.data, &rampC)
-}
