+++ monitor.c
@@ -15,3 +15,11 @@ void glfwMonitorCB(GLFWmonitor* monitor, int event) {
 void glfwSetMonitorCallbackCB() {
 	glfwSetMonitorCallback(glfwMonitorCB);
 }
+
+unsigned int GetGammaAtIndex(unsigned short *color, int i) {
+	return color[i];
+}
+
+void SetGammaAtIndex(unsigned short *color, int i, unsigned short value) {
+	color[i] = value;
+}
